requirements.txt
# LangChain core packages
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Vector store
faiss-cpu>=1.7.4

# YouTube transcript
youtube-transcript-api>=0.6.2

# Web interface
streamlit>=1.38.0

# Environment variables
python-dotenv>=1.0.0

# Google Generative AI
google-generativeai>=0.8.0
